+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Moved NC And IA gage pages links inside StreamStats gage page modal
 - Watershed symbols in the legend
 - StreamStats Gage Page link in gage popups to a button
+- Updated to Google Analytics 4
 
 ### Deprecated 
 
@@ -24,8 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed 
 
 - Legacy NWIS link for gage pages
-- Gulp
 - NWIS page link in gage popups
+- Angulartics
   
 ### Fixed